feat: add prompt input
